+++ src/Entidades/Entidade.cpp
@@ -2,7 +2,7 @@
 
 namespace Entidades {
 
-Entidade::Entidade() : Ente(), x(0), y(0) {}
+Entidade::Entidade() : Ente(), x(0.0f), y(0.0f) {}
 
 Entidade::~Entidade() {}
 
@@ -10,11 +10,11 @@ void Entidade::salvarDataBuffer() {
     
 }
 
-int Entidade::getX() const {
+float Entidade::getX() const {
     return x;
 }
 
-int Entidade::getY() const {
+float Entidade::getY() const {
     return y;
 }
